Re-worked includes. Cleaned .pro (it has Mn & Ex files, and a weird order), etc.

+++ src/streammanager.cpp
@@ -1,14 +1,12 @@
 #include "streammanager.h"
 #include <flexsea.h>
-#include <flexsea_comm.h>
 #include <flexsea_system.h>
-#include <flexsea_cmd_data.h>
-#include <flexsea_cmd_in_control.h>
-#include <flexsea_cmd_stream.h>
-#include "../flexsea-user/inc/flexsea_cmd_user.h"
 #include <flexsea_board.h>
 #include <w_event.h>
 #include <QDebug>
+#include <cmd-MIT_2DoF_Ankle_v1.h>
+#include <cmd-RICNU_Knee_v1.h>
+#include <cmd-MotorTestBench.h>
 
 StreamManager::StreamManager(QObject *parent, SerialDriver* driver) :
 	QObject(parent),
@@ -292,7 +290,6 @@ void StreamManager::sendCommands(int index)
 	}
 }
 
-
 void StreamManager::sendCommandReadAll(uint8_t slaveId)
 {
 	//1) Stream
@@ -305,7 +302,7 @@ void StreamManager::sendCommandReadAllRicnu(uint8_t slaveId)
 	(void) slaveId;
 	if(ricnuOffsets.size() < 1) return;
 	static int index = 0;
-	tx_cmd_ricnu_r(TX_N_DEFAULT, ricnuOffsets.at(index));
+	tx_cmd_ricnu_r(TX_N_DEFAULT, (uint8_t)(ricnuOffsets.at(index)));
 	index++;
 	index %= ricnuOffsets.size();
 	tryPackAndSend(CMD_READ_ALL_RICNU, slaveId);
